% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cubic_spline}
\alias{cubic_spline}
\title{Find Function using Cubic Spline Interpolation}
\usage{
cubic_spline(x, nbreaks = 10)
}
\arguments{
\item{x}{The variable to perform the spline over}

\item{nbreaks}{The number of breaks for the cumulative sum of frequencies applied to x}
}
\value{
a cubic spline interpolation function of x
}
\description{
This performs cubic spline interpolation over the cumulative
sum of frequencies of a variable.
}
