% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{find_from_theta}
\alias{find_from_theta}
\title{Approximate the function}
\usage{
find_from_theta(x, y, vals)
}
\arguments{
\item{x}{The first variable}

\item{y}{The second variable}

\item{vals}{the values to calculate their points on the line}
}
\description{
Approximate the funtion between two variables and calculate 
points from the approximated function
}
