% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming_utilities.R
\name{scale_color_gradient_oberd}
\alias{scale_color_gradient_oberd}
\title{Default continuous geom colors for oberd}
\usage{
scale_color_gradient_oberd(theme = "dark", o_key = list(dark = c("#165E85",
  "#DBF733"), light = c("#0093CA", "#DBF733")))
}
\arguments{
\item{theme}{the colors scheme to go with the theme}

\item{o_key}{list of different color schemes}
}
\description{
Default continuous geom colors for oberd
}
