% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming_utilities.R
\name{scale_color_oberd}
\alias{scale_color_oberd}
\title{Default discrete geom colors for oberd}
\usage{
scale_color_oberd(theme = "dark", o_key = list(dark = c("#165E85",
  "#94979D", "#161D27", "#5B677F"), light = c("#0093CA", "#DBF733", "#FFFFFF",
  "#94979D")))
}
\arguments{
\item{theme}{the colors scheme to go with the theme}

\item{o_key}{list of different color schemes}
}
\description{
Default discrete geom colors for oberd
}
