% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{spline_adjustment}
\alias{spline_adjustment}
\title{Adjust the splines derivative}
\usage{
spline_adjustment(x, deriv)
}
\arguments{
\item{x}{The variable to plot the derivative against}

\item{deriv}{The derivative of the cubic spline}
}
\value{
the adjusted derivative
}
\description{
This function adjusts the derivative of a cubic spline so that
when plotted against another variable, the AUC is equal to 1
}
