% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cubic_spline}
\alias{cubic_spline}
\title{Find Function using Cubic Spline Interpolation}
\usage{
cubic_spline(x, break_size = 10, min = 0, max = 100)
}
\arguments{
\item{x}{The variable to perform the spline over}

\item{break_size}{The width of each break for the cumulative sum of frequencies applied to x}

\item{min}{The minimum value for the break sequence}

\item{max}{The maximum value for the break sequence}
}
\value{
a cubic spline interpolation function of x
}
\description{
This performs cubic spline interpolation over the cumulative
sum of frequencies of a variable.
}
